+++ aws_account/03_apps/_providers.tf
@@ -10,7 +10,7 @@ terraform {
     # https://registry.terraform.io/providers/hashicorp/aws/latest
     aws = {
       source  = "hashicorp/aws"
-      version = "5.90.0"
+      version = "5.94.1"
     }
   }
   backend "s3" {